requirements.txt
python-telegram-bot==20.7
openai>=1.0.0
requests>=2.31.0
pandas>=2.1.0
matplotlib>=3.8.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
aiosqlite
pydantic>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
scipy>=1.11.0